mario3

+++ SysVentas/src/main/java/pe/edu/upeu/sysventas/service/ICompraDetalleService.java
@@ -2,5 +2,5 @@ package pe.edu.upeu.sysventas.service;
 
 import pe.edu.upeu.sysventas.model.CompraDetalle;
 
-public interface ICompraDetalleService extends ICrudGenericService<CompraDetalle,Integer> {
+public interface ICompraDetalleService extends ICrudGenericService<CompraDetalle,Long>{
 }
